--- a/tools/training-benchmark.py
+++ b/tools/training-benchmark.py
@@ -166,12 +166,12 @@ if __name__ == "__main__":
             atomic_properties = []
 
             for chunk_species, chunk_coordinates in batch_x:
-                atomic_chunk = {'species': chunk_species, 'coordinates': chunk_coordinates}
+                atomic_chunk = {'species': chunk_species.to(parser.device), 'coordinates': chunk_coordinates.to(parser.device)}
                 atomic_properties.append(atomic_chunk)
                 num_atoms.append((chunk_species >= 0).to(true_energies.dtype).sum(dim=1))
 
             atomic_properties = torchani.utils.pad_atomic_properties(atomic_properties)
-            predicted_energies = model((atomic_properties['species'], atomic_properties['coordinates'])).energies
+            predicted_energies = model((atomic_properties['species'], atomic_properties['coordinates'])).energies.to(true_energies.dtype)
 
             num_atoms = torch.cat(num_atoms)
             loss = (mse(predicted_energies, true_energies) / num_atoms.sqrt()).mean()