Massive training time improvement by doing a simple substitution in nnp_training