@farhadrgh 2020-03-19
I think the issue is shuffled dataset is passing chunks in CPU device. You will need to:
```
chunk_species = chunk_species.to(parser.device)
chunk_coordinates = chunk_coordinates.to(parser.device)
```